requirements.txt
amqp==2.6.1
backcall==0.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==2.2.16
django-redis==4.12.1
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kombu==4.6.11
mysqlclient==2.0.1
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.7
Pygments==2.7.1
pytz==2020.1
qiniu==7.2.10
redis==3.5.3
requests==2.24.0
sqlparse==0.3.1
tencentcloud-sdk-python==3.0.261
traitlets==5.0.4
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zipp==3.1.0
